+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii EAuth Change Log
 * Update Twitter API to 1.1 (fixed #61)
 * Add User-Agent header in GitHubOAuthService.php (fixed #60)
 * Reformat source code
+* Added ability to change OpenID realm
 
 ### Changes before 15.06.2013
 * Added Microsoft Live OAuth provider.